+++ src/impl_prim.rs
@@ -280,8 +280,6 @@ impl_primitive!(
     (2(0), TestAssert, Impure),
     (2, ValidateImpl(Option<usize>, Option<SubSide>)),
     (1, MvImpl(MvMode)),
-    (2, InnerProduct),
-    (2, OuterProduct),
     (2, RegressiveProduct),
     /// Validate that a non-boxed variant field has a valid type and rank
     (1, ValidateNonBoxedVariant),